clusterization final version, code style upgrade, architecture improvement
