+++ Ants/Screens/ANTS_OBJORA.html
@@ -2,7 +2,7 @@
 <head>
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Ant Colony Optimization</title>
-    <link rel="stylesheet" href="../Styles/Ants.css">
+    <link rel="stylesheet" href="../Styles/ants.css">
 </head>
 <body>
 
@@ -53,6 +53,6 @@
     <div class="bottom"></div>
 </div>
 <div class="side-border-right"></div>
-<script src="../Scripts/Ants.js"></script>
+<script src="../Scripts/ants.js"></script>
 </body>
 </html>
