+++ Ants/Screens/ants.html
@@ -33,7 +33,7 @@
         <button onclick="" id="wallBtn" class = "wall-btn"></button>
         <button onclick="" id="foodBtn" class = "food-btn"></button>
         <div class="text-field">
-            <input class="text-field__input" type="text" pattern="\d*" name="ants" id="ants" placeholder="ants" value="500">
+            <input class="text-field-input-ants" type="text" pattern="\d*" name="ants" id="ants" placeholder="ants" value="500">
         </div>
         <button onclick="" id="clearBtn" class = "remove-btn"></button>
         <button onclick="" id="nestBtn" class = "nest-btn"></button>
