+++ Clusterization/Screens/clusterization.html
@@ -1,10 +1,10 @@
-
-<div style="display: inline-flex">
+<div class="algorithm-main-part">
     <h1 class="header-text">
         CLUSTER <br>
         ANALYSIS
         <div class="underline"></div>
     </h1>
+    <img src="Clusterization/images/iconСluster.png" alt="" class="algorithm-symbol">
 </div>
 <div style="text-align: left">
     <h2 class="interview-right">Clustering analysis is a powerful method in data mining that automatically groups
