+++ Clusterization/Scripts/clusterization.js
@@ -1,13 +1,15 @@
-const WIDTH = 25;
-const HEIGHT = 25;
-const kMeanStartBtn = document.getElementById('kmeans-start-btn');
-const hierarchicalStartBtn = document.getElementById('hierarchical-start-btn');
-const miniBatchKMeansStartBtn = document.getElementById('miniBatchKMeans-start-btn');
-const clusterInput = document.getElementById('clusters');
+import {Algorithms} from "./algorithms.js";
+import {Config} from "./config.js";
+
+
+const kMeanStartBtn = document.getElementById(Config.DOM_IDS.KMEANS_BTN);
+const hierarchicalStartBtn = document.getElementById(Config.DOM_IDS.HIERARCHICAL_BTN);
+const miniBatchKMeansStartBtn = document.getElementById(Config.DOM_IDS.MINIBATCH_BTN);
+const clusterInput = document.getElementById(Config.DOM_IDS.CLUSTERS_INPUT);
+const clearBtn = document.getElementById(Config.DOM_IDS.CLEAR_BTN);
 
 let cells = [];
 let chosenCells = new Set();
-let clusterCntInputField = document.getElementById('clusters');
 let clusterCnt = 0;
 let pointsArray = [];
 let kMeansClusters;
@@ -20,44 +22,6 @@ const screenState = {
     thirdAlg: false,
 };
 
-let algorithmColors = [
-    [
-        '#FF0000',
-        '#33FF57',
-        '#3357FF',
-        '#F3FF33',
-        '#FF33F3',
-        '#33FFF3',
-        '#FF8C33',
-        '#9933FF',
-        '#33FF99',
-    ],
-
-    [
-        '#FF9999',
-        '#99FF99',
-        '#9999FF',
-        '#FFFF99',
-        '#FF99FF',
-        '#99FFFF',
-        '#FFCC99',
-        '#CC99FF',
-        '#99FFCC',
-    ],
-
-    [
-        '#CC8888',
-        '#88CC88',
-        '#8888CC',
-        '#CCCC88',
-        '#CC88CC',
-        '#88CCCC',
-        '#CC9988',
-        '#AA88CC',
-        '#88CCAA',
-    ],
-];
-
 class Location {
     constructor(y, x) {
         this.y = y;
@@ -65,263 +29,21 @@ class Location {
     }
 }
 
-function kMeansClustering(points, k, maxIterations = 100, convergenceThreshold = 0.001) {
-    if (points.length < k) {
-        return Array(points.length)
-            .fill()
-            .map((_, i) => i);
-    }
-
-    let centroids = [{ ...points[Math.floor(Math.random() * points.length)] }];
-
-    while (centroids.length < k) {
-        let distances = points.map((point) => {
-            return Math.min(
-                ...centroids.map(
-                    (centroid) =>
-                        Math.pow(point.x - centroid.x, 2) + Math.pow(point.y - centroid.y, 2)
-                )
-            );
-        });
-
-        let sum = distances.reduce((a, b) => a + b, 0);
-        let random = Math.random() * sum;
-
-        let acc = 0;
-        for (let i = 0; i < points.length; i++) {
-            acc += distances[i];
-            if (acc >= random) {
-                centroids.push({ ...points[i] });
-                break;
-            }
-        }
-    }
-
-    let clusterAssignments = Array(points.length);
-    let iterations = 0;
-
-    while (iterations < maxIterations) {
-        for (let i = 0; i < points.length; i++) {
-            const point = points[i];
-            let minDistance = Infinity;
-            let closestCluster = 0;
-
-            for (let j = 0; j < centroids.length; j++) {
-                const distance = Math.sqrt(
-                    Math.pow(point.y - centroids[j].y, 2) + Math.pow(point.x - centroids[j].x, 2)
-                );
-
-                if (distance < minDistance) {
-                    minDistance = distance;
-                    closestCluster = j;
-                }
-            }
-            clusterAssignments[i] = closestCluster;
-        }
-
-        let newCentroids = Array(k)
-            .fill()
-            .map(() => ({ y: 0, x: 0, count: 0 }));
-
-        for (let i = 0; i < points.length; i++) {
-            const clusterIndex = clusterAssignments[i];
-            newCentroids[clusterIndex].y += points[i].y;
-            newCentroids[clusterIndex].x += points[i].x;
-            newCentroids[clusterIndex].count++;
-        }
-
-        let maxCentroidShift = 0;
-        for (let i = 0; i < k; i++) {
-            if (newCentroids[i].count > 0) {
-                const newY = newCentroids[i].y / newCentroids[i].count;
-                const newX = newCentroids[i].x / newCentroids[i].count;
-
-                const shift = Math.sqrt(
-                    Math.pow(newY - centroids[i].y, 2) + Math.pow(newX - centroids[i].x, 2)
-                );
-                maxCentroidShift = Math.max(maxCentroidShift, shift);
-
-                centroids[i] = {
-                    y: newY,
-                    x: newX,
-                };
-            }
-        }
-
-        iterations++;
-
-        if (maxCentroidShift < convergenceThreshold) {
-            break;
-        }
-    }
-    let totalDistance = 0;
-    for (let i = 0; i < points.length; i++) {
-        const centroid = centroids[clusterAssignments[i]];
-        totalDistance += Math.sqrt(
-            Math.pow(points[i].y - centroid.y, 2) + Math.pow(points[i].x - centroid.x, 2)
-        );
-    }
-
-    return clusterAssignments;
-}
-
-function miniBatchKMeansClustering(points, k) {
-    const batchSize = Math.min(10, points.length);
-    const maxIterations = 50;
-
-    let centroids = [];
-    let usedIndices = new Set();
-
-    while (centroids.length < k) {
-        const randomIndex = Math.floor(Math.random() * points.length);
-        if (!usedIndices.has(randomIndex)) {
-            usedIndices.add(randomIndex);
-            centroids.push({
-                x: points[randomIndex].x,
-                y: points[randomIndex].y,
-                count: 0,
-            });
-        }
-    }
-
-    let clusterAssignments = new Array(points.length).fill(0);
-
-    for (let iteration = 0; iteration < maxIterations; iteration++) {
-        let batchIndices = new Set();
-        while (batchIndices.size < batchSize) {
-            batchIndices.add(Math.floor(Math.random() * points.length));
-        }
-        batchIndices = Array.from(batchIndices);
-
-        batchIndices.forEach((pointIndex) => {
-            const point = points[pointIndex];
-            let minDistance = Infinity;
-            let nearestCentroid = 0;
-
-            centroids.forEach((centroid, centroidIndex) => {
-                const distance = calculateDistance(point, centroid);
-                if (distance < minDistance) {
-                    minDistance = distance;
-                    nearestCentroid = centroidIndex;
-                }
-            });
-
-            clusterAssignments[pointIndex] = nearestCentroid;
-
-            const centroid = centroids[nearestCentroid];
-            centroid.count += 1;
-            const learningRate = 1 / centroid.count;
-
-            centroid.x = centroid.x * (1 - learningRate) + point.x * learningRate;
-            centroid.y = centroid.y * (1 - learningRate) + point.y * learningRate;
-        });
-    }
-
-    return clusterAssignments;
-}
-
-function hierarchicalClustering(points, k) {
-    let clusters = points.map((point, index) => ({
-        points: [point],
-        index: index,
-    }));
-
-    let distances = [];
-    for (let i = 0; i < clusters.length; i++) {
-        for (let j = i + 1; j < clusters.length; j++) {
-            distances.push({
-                from: i,
-                to: j,
-                distance: calculateDistance(clusters[i].points[0], clusters[j].points[0]),
-            });
-        }
-    }
-
-    while (clusters.filter((c) => c !== null).length > k) {
-        let minDistance = Infinity;
-        let mergeIndexes = null;
-
-        distances.forEach((d) => {
-            if (d.distance < minDistance && clusters[d.from] !== null && clusters[d.to] !== null) {
-                minDistance = d.distance;
-                mergeIndexes = [d.from, d.to];
-            }
-        });
-
-        if (!mergeIndexes) break;
-
-        const [from, to] = mergeIndexes;
-        clusters[from].points = clusters[from].points.concat(clusters[to].points);
-        clusters[to] = null;
-
-        distances = distances.map((d) => {
-            if (d.from === to || d.to === to) {
-                return { ...d, distance: Infinity };
-            }
-            if (d.from === from || d.to === from) {
-                const otherIndex = d.from === from ? d.to : d.from;
-                if (clusters[otherIndex] === null) return d;
-
-                return {
-                    ...d,
-                    distance: calculateAverageDistance(
-                        clusters[from].points,
-                        clusters[otherIndex].points
-                    ),
-                };
-            }
-            return d;
-        });
-    }
-
-    let clusterAssignments = new Array(points.length);
-    let clusterNumber = 0;
-
-    clusters.forEach((cluster) => {
-        if (cluster === null) return;
-
-        cluster.points.forEach((point) => {
-            const pointIndex = points.findIndex((p) => p.x === point.x && p.y === point.y);
-            clusterAssignments[pointIndex] = clusterNumber;
-        });
-        clusterNumber++;
-    });
-
-    return clusterAssignments;
-}
-
-function calculateDistance(point1, point2) {
-    return Math.sqrt(Math.pow(point1.x - point2.x, 2) + Math.pow(point1.y - point2.y, 2));
-}
-
-function calculateAverageDistance(points1, points2) {
-    let totalDistance = 0;
-    let count = 0;
-
-    points1.forEach((p1) => {
-        points2.forEach((p2) => {
-            totalDistance += calculateDistance(p1, p2);
-            count++;
-        });
-    });
-
-    return totalDistance / count;
-}
 
 function showKMeansClusters() {
-    if (clusterCntInputField.value === '') {
+    if (clusterInput.value === '') {
         return;
     }
 
     if (screenState.firstAlg === false) {
-        kMeanStartBtn.classList.remove('kmeans-start-btn-deactivate');
-        kMeanStartBtn.classList.add('kmeans-start-btn-activate');
+        kMeanStartBtn.classList.remove(Config.CSS_CLASSES.KMEANS_INACTIVE);
+        kMeanStartBtn.classList.add(Config.CSS_CLASSES.KMEANS_ACTIVE);
         screenState.firstAlg = true;
         applyKMeansColoring();
     } else {
         screenState.firstAlg = false;
-        kMeanStartBtn.classList.remove('kmeans-start-btn-activate');
-        kMeanStartBtn.classList.add('kmeans-start-btn-deactivate');
+        kMeanStartBtn.classList.remove(Config.CSS_CLASSES.KMEANS_ACTIVE);
+        kMeanStartBtn.classList.add(Config.CSS_CLASSES.KMEANS_INACTIVE);
         makeAllTransparent();
 
         if (screenState.secondAlg === true) {
@@ -338,18 +60,18 @@ function showKMeansClusters() {
 }
 
 function showHierarchicalClusters() {
-    if (clusterCntInputField.value === '') {
+    if (clusterInput.value === '') {
         return;
     }
 
     if (screenState.secondAlg === false) {
-        hierarchicalStartBtn.classList.remove('hierarchical-start-btn-deactivate');
-        hierarchicalStartBtn.classList.add('hierarchical-start-btn-activate');
+        hierarchicalStartBtn.classList.remove(Config.CSS_CLASSES.HIERARCHICAL_INACTIVE);
+        hierarchicalStartBtn.classList.add(Config.CSS_CLASSES.HIERARCHICAL_ACTIVE);
         screenState.secondAlg = true;
         applyHierarchicalColoring();
     } else {
-        hierarchicalStartBtn.classList.add('hierarchical-start-btn-deactivate');
-        hierarchicalStartBtn.classList.remove('hierarchical-start-btn-activate');
+        hierarchicalStartBtn.classList.add(Config.CSS_CLASSES.HIERARCHICAL_INACTIVE);
+        hierarchicalStartBtn.classList.remove(Config.CSS_CLASSES.HIERARCHICAL_ACTIVE);
         screenState.secondAlg = false;
 
         makeAllTransparent();
@@ -368,18 +90,18 @@ function showHierarchicalClusters() {
 }
 
 function showMiniBatchKMeansClusters() {
-    if (clusterCntInputField.value === '') {
+    if (clusterInput.value === '') {
         return;
     }
 
     if (screenState.thirdAlg === false) {
-        miniBatchKMeansStartBtn.classList.remove('miniBatchKMeans-start-btn-deactivate');
-        miniBatchKMeansStartBtn.classList.add('miniBatchKMeans-start-btn-activate');
+        miniBatchKMeansStartBtn.classList.remove(Config.CSS_CLASSES.MINI_BATCH_INACTIVE);
+        miniBatchKMeansStartBtn.classList.add(Config.CSS_CLASSES.MINI_BATCH_ACTIVE);
         screenState.thirdAlg = true;
         applyMiniBatchKMeansColoring();
     } else {
-        miniBatchKMeansStartBtn.classList.add('miniBatchKMeans-start-btn-deactivate');
-        miniBatchKMeansStartBtn.classList.remove('miniBatchKMeans-start-btn-activate');
+        miniBatchKMeansStartBtn.classList.add(Config.CSS_CLASSES.MINI_BATCH_INACTIVE);
+        miniBatchKMeansStartBtn.classList.remove(Config.CSS_CLASSES.MINI_BATCH_ACTIVE);
         screenState.thirdAlg = false;
 
         makeAllTransparent();
@@ -401,136 +123,141 @@ function cancelAllAlg() {
     pointsArray = Array.from(chosenCells);
     pointsArray.forEach((cell) => {
         const cellElement = document.getElementById(
-            `algorithm-clusterize-cell-${cell.y}-${cell.x}`
+            `${Config.CSS_CLASSES.GRID_CELL}-${cell.y}-${cell.x}`
         );
         for (let i = 0; i < 3; i++) {
-            cellElement.children[i].style.backgroundColor = 'transparent';
+            cellElement.children[i].style.backgroundColor = Config.COLORS.TRANSPARENT;
             if (i === 1) {
                 for (let j = 0; j < 2; j++) {
-                    cellElement.children[1].children[j].style.backgroundColor = 'transparent';
+                    cellElement.children[1].children[j].style.backgroundColor = Config.COLORS.TRANSPARENT;
                 }
             }
         }
-        cellElement.style.backgroundColor = '#b5b094';
+        cellElement.style.backgroundColor = Config.COLORS.CHOSEN_CELL;
     });
 }
 
 function makeAllTransparent() {
     pointsArray.forEach((cell) => {
         const cellElement = document.getElementById(
-            `algorithm-clusterize-cell-${cell.y}-${cell.x}`
+            `${Config.CSS_CLASSES.GRID_CELL}-${cell.y}-${cell.x}`
         );
-        cellElement.style.backgroundColor = 'transparent';
-        cellElement.children[0].style.backgroundColor = 'transparent';
-        cellElement.children[1].style.backgroundColor = 'transparent';
-        cellElement.children[2].style.backgroundColor = 'transparent';
-        cellElement.children[1].children[0].style.backgroundColor = 'transparent';
-        cellElement.children[1].children[1].style.backgroundColor = 'transparent';
+        cellElement.style.backgroundColor = Config.COLORS.TRANSPARENT;
+        cellElement.children[0].style.backgroundColor = Config.COLORS.TRANSPARENT;
+        cellElement.children[1].style.backgroundColor = Config.COLORS.TRANSPARENT;
+        cellElement.children[2].style.backgroundColor = Config.COLORS.TRANSPARENT;
+        cellElement.children[1].children[0].style.backgroundColor = Config.COLORS.TRANSPARENT;
+        cellElement.children[1].children[1].style.backgroundColor = Config.COLORS.TRANSPARENT;
     });
 }
 
 function applyKMeansColoring() {
-    console.log(screenState);
     pointsArray = Array.from(chosenCells);
-    clusterCnt = parseInt(clusterCntInputField.value, 10);
-    kMeansClusters = kMeansClustering(pointsArray, clusterCnt);
+    clusterCnt = parseInt(clusterInput.value, 10);
+    kMeansClusters = Algorithms.kMeansClustering(pointsArray, Math.min(clusterCnt, pointsArray.length));
     pointsArray.forEach((cell, index) => {
         const group = kMeansClusters[index];
         const cellElement = document.getElementById(
-            `algorithm-clusterize-cell-${cell.y}-${cell.x}`
+            `${Config.CSS_CLASSES.GRID_CELL}-${cell.y}-${cell.x}`
         );
         if (screenState.secondAlg === true && screenState.thirdAlg === true) {
-            cellElement.children[0].style.backgroundColor = algorithmColors[0][group];
+            cellElement.children[0].style.backgroundColor = Config.COLORS.ALGORITHM_COLORS[0][group];
             applyHierarchicalColoring();
         } else if (
             (screenState.secondAlg === false && screenState.thirdAlg === true) ||
             (screenState.secondAlg === true && screenState.thirdAlg === false)
         ) {
-            cellElement.children[0].style.backgroundColor = algorithmColors[0][group];
-            cellElement.children[1].children[0].style.backgroundColor = algorithmColors[0][group];
+            cellElement.children[0].style.backgroundColor = Config.COLORS.ALGORITHM_COLORS[0][group];
+            cellElement.children[1].children[0].style.backgroundColor = Config.COLORS.ALGORITHM_COLORS[0][group];
         } else if (screenState.secondAlg === false && screenState.thirdAlg === false) {
-            cellElement.style.backgroundColor = algorithmColors[0][group];
+            cellElement.style.backgroundColor = Config.COLORS.ALGORITHM_COLORS[0][group];
         }
     });
 }
 
 function applyMiniBatchKMeansColoring() {
-    console.log(screenState);
     pointsArray = Array.from(chosenCells);
-    clusterCnt = parseInt(clusterCntInputField.value, 10);
-    miniBatchKMeansClusters = miniBatchKMeansClustering(pointsArray, clusterCnt);
+    clusterCnt = parseInt(clusterInput.value, 10);
+    miniBatchKMeansClusters = Algorithms.miniBatchKMeansClustering(pointsArray, Math.min(clusterCnt, pointsArray.length));
     pointsArray.forEach((cell, index) => {
         const group = miniBatchKMeansClusters[index];
         const cellElement = document.getElementById(
-            `algorithm-clusterize-cell-${cell.y}-${cell.x}`
+            `${Config.CSS_CLASSES.GRID_CELL}-${cell.y}-${cell.x}`
         );
         if (screenState.secondAlg === true && screenState.firstAlg === true) {
-            cellElement.children[2].style.backgroundColor = algorithmColors[2][group];
+            cellElement.children[2].style.backgroundColor = Config.COLORS.ALGORITHM_COLORS[2][group];
             applyHierarchicalColoring();
         } else if (
             (screenState.secondAlg === false && screenState.firstAlg === true) ||
             (screenState.secondAlg === true && screenState.firstAlg === false)
         ) {
-            cellElement.children[2].style.backgroundColor = algorithmColors[2][group];
-            cellElement.children[1].children[1].style.backgroundColor = algorithmColors[2][group];
+            cellElement.children[2].style.backgroundColor = Config.COLORS.ALGORITHM_COLORS[2][group];
+            cellElement.children[1].children[1].style.backgroundColor = Config.COLORS.ALGORITHM_COLORS[2][group];
         } else if (screenState.secondAlg === false && screenState.firstAlg === false) {
-            cellElement.style.backgroundColor = algorithmColors[2][group];
+            cellElement.style.backgroundColor = Config.COLORS.ALGORITHM_COLORS[2][group];
         }
     });
 }
 
 function applyHierarchicalColoring() {
     pointsArray = Array.from(chosenCells);
-    clusterCnt = parseInt(clusterCntInputField.value, 10);
-    hierarchicalClusters = hierarchicalClustering(pointsArray, clusterCnt);
+    clusterCnt = parseInt(clusterInput.value, 10);
+    hierarchicalClusters = Algorithms.hierarchicalClustering(pointsArray, Math.min(clusterCnt, pointsArray.length));
     pointsArray.forEach((cell, index) => {
         const group = hierarchicalClusters[index];
         const cellElement = document.getElementById(
-            `algorithm-clusterize-cell-${cell.y}-${cell.x}`
+            `${Config.CSS_CLASSES.GRID_CELL}-${cell.y}-${cell.x}`
         );
         if (screenState.firstAlg === true && screenState.thirdAlg === true) {
-            cellElement.children[1].children[0].style.backgroundColor = algorithmColors[1][group];
-            cellElement.children[1].children[1].style.backgroundColor = algorithmColors[1][group];
+            cellElement.children[1].children[0].style.backgroundColor = Config.COLORS.ALGORITHM_COLORS[1][group];
+            cellElement.children[1].children[1].style.backgroundColor = Config.COLORS.ALGORITHM_COLORS[1][group];
         } else if (screenState.firstAlg === true && screenState.thirdAlg === false) {
-            cellElement.children[2].style.backgroundColor = algorithmColors[1][group];
-            cellElement.children[1].children[1].style.backgroundColor = algorithmColors[1][group];
+            cellElement.children[2].style.backgroundColor = Config.COLORS.ALGORITHM_COLORS[1][group];
+            cellElement.children[1].children[1].style.backgroundColor = Config.COLORS.ALGORITHM_COLORS[1][group];
         } else if (screenState.firstAlg === false && screenState.thirdAlg === true) {
-            cellElement.children[0].style.backgroundColor = algorithmColors[1][group];
-            cellElement.children[1].children[0].style.backgroundColor = algorithmColors[1][group];
+            cellElement.children[0].style.backgroundColor = Config.COLORS.ALGORITHM_COLORS[1][group];
+            cellElement.children[1].children[0].style.backgroundColor = Config.COLORS.ALGORITHM_COLORS[1][group];
         } else if (screenState.firstAlg === false && screenState.thirdAlg === false) {
-            cellElement.style.backgroundColor = algorithmColors[1][group];
+            cellElement.style.backgroundColor = Config.COLORS.ALGORITHM_COLORS[1][group];
         }
     });
 }
 
 function initGrid() {
-    let grid = document.getElementById('algorithm-clusterize-grid');
+    let grid = document.getElementById(Config.DOM_IDS.GRID);
     grid.innerHTML = '';
-    cells = Array(WIDTH)
+    cells = Array(Config.GRID.WIDTH)
         .fill()
-        .map(() => Array(HEIGHT));
+        .map(() => Array(Config.GRID.HEIGHT));
 
-    for (let i = 0; i < WIDTH; i++) {
+    for (let i = 0; i < Config.GRID.WIDTH; i++) {
         const row = document.createElement('div');
-        row.className = 'algorithm-clusterize-row';
-        for (let j = 0; j < HEIGHT; j++) {
+        row.className = Config.CSS_CLASSES.GRID_ROW;
+        for (let j = 0; j < Config.GRID.HEIGHT; j++) {
             const cell = document.createElement('div');
-            cell.className = 'algorithm-clusterize-cell';
-            cell.id = `algorithm-clusterize-cell-${i}-${j}`;
+            cell.className = Config.CSS_CLASSES.GRID_CELL;
+            cell.id = `${Config.CSS_CLASSES.GRID_CELL}-${i}-${j}`;
             cells[i][j] = new Location(i, j);
             cell.addEventListener('click', function () {
                 if (chosenCells.has(cells[i][j])) {
-                    document.getElementById(
-                        `algorithm-clusterize-cell-${i}-${j}`
-                    ).style.backgroundColor = 'transparent';
+                    const currentCell = document.getElementById(
+                        `${Config.CSS_CLASSES.GRID_CELL}-${i}-${j}`
+                    )
+                    currentCell.style.backgroundColor = Config.COLORS.TRANSPARENT;
+                    currentCell.children[0].style.backgroundColor = Config.COLORS.TRANSPARENT;
+                    currentCell.children[1].style.backgroundColor = Config.COLORS.TRANSPARENT;
+                    currentCell.children[2].style.backgroundColor = Config.COLORS.TRANSPARENT;
+                    currentCell.children[1].children[0].style.backgroundColor = Config.COLORS.TRANSPARENT;
+                    currentCell.children[1].children[1].style.backgroundColor = Config.COLORS.TRANSPARENT;
+
                     chosenCells.delete(cells[i][j]);
                 } else {
                     document.getElementById(
-                        `algorithm-clusterize-cell-${i}-${j}`
-                    ).style.backgroundColor = '#b5b094';
+                        `${Config.CSS_CLASSES.GRID_CELL}-${i}-${j}`
+                    ).style.backgroundColor = Config.COLORS.CHOSEN_CELL;
                     chosenCells.add(cells[i][j]);
                 }
-                if (clusterCntInputField.value !== '') {
+                if (clusterInput.value !== '') {
                     if (screenState.firstAlg === true) {
                         applyKMeansColoring();
                     }
@@ -545,11 +272,11 @@ function initGrid() {
 
             for (let i = 0; i < 3; i++) {
                 let subCell = document.createElement('div');
-                subCell.className = 'algorithm-clusterize-sub-cell';
+                subCell.className = Config.CSS_CLASSES.GRID_SUBCELL;
                 if (i === 1) {
                     for (let j = 0; j < 2; j++) {
                         let subSubCell = document.createElement('div');
-                        subSubCell.className = 'algorithm-clusterize-sub-sub-cell';
+                        subSubCell.className = Config.CSS_CLASSES.GRID_SUBSUBCELL;
                         subCell.appendChild(subSubCell);
                     }
                 }
@@ -566,22 +293,28 @@ function clearGrid() {
     chosenCells.clear();
 
     screenState.firstAlg = false;
-    kMeanStartBtn.classList.add('kmeans-start-btn-deactivate');
-    kMeanStartBtn.classList.remove('kmeans-start-btn-activate');
+    kMeanStartBtn.classList.add(Config.CSS_CLASSES.KMEANS_INACTIVE);
+    kMeanStartBtn.classList.remove(Config.CSS_CLASSES.KMEANS_ACTIVE);
 
     screenState.secondAlg = false;
-    hierarchicalStartBtn.classList.add('hierarchical-start-btn-deactivate');
-    hierarchicalStartBtn.classList.remove('hierarchical-start-btn-activate');
+    hierarchicalStartBtn.classList.add(Config.CSS_CLASSES.HIERARCHICAL_INACTIVE);
+    hierarchicalStartBtn.classList.remove(Config.CSS_CLASSES.HIERARCHICAL_ACTIVE);
 
     screenState.thirdAlg = false;
-    miniBatchKMeansStartBtn.classList.add('miniBatchKMeans-start-btn-deactivate');
-    miniBatchKMeansStartBtn.classList.remove('miniBatchKMeans-start-btn-activate');
+    miniBatchKMeansStartBtn.classList.add(Config.CSS_CLASSES.MINI_BATCH_INACTIVE);
+    miniBatchKMeansStartBtn.classList.remove(Config.CSS_CLASSES.MINI_BATCH_ACTIVE);
 }
+
 
 kMeanStartBtn.addEventListener('click', showKMeansClusters);
 hierarchicalStartBtn.addEventListener('click', showHierarchicalClusters);
 miniBatchKMeansStartBtn.addEventListener('click', showMiniBatchKMeansClusters);
-
+clusterInput.addEventListener('keypress', function (e) {
+    if (e.key === '.' || e.key === ',') {
+        e.preventDefault();
+    }
+});
+clearBtn.addEventListener('click', clearGrid);
 clusterInput.addEventListener('input', function () {
     this.value = this.value.replace(/[^1-9]/g, '');
 
@@ -590,8 +323,8 @@ clusterInput.addEventListener('input', function () {
     }
 
     const num = parseInt(this.value);
-    if (num > 9) {
-        this.value = '9';
+    if (num > Config.CLUSTERING.MAX_CLUSTERS) {
+        this.value = Config.CLUSTERING.MAX_CLUSTERS;
     }
 
     if (parseInt(this.value) !== clusterCnt && this.value !== '') {
@@ -607,12 +340,8 @@ clusterInput.addEventListener('input', function () {
     }
 });
 
-clusterInput.addEventListener('keypress', function (e) {
-    if (e.key === '.' || e.key === ',') {
-        e.preventDefault();
-    }
-});
+
 
 initGrid();
 
-document.getElementById('clear-Btn').addEventListener('click', clearGrid);
+
