+++ Clusterization/Styles/clusterization.css
@@ -40,7 +40,7 @@
 }
 
 #kmeans-start-btn {
-    margin-top: 100px;
+    margin-top: 10px;
     width: 60px;
     height: 60px;
 }
@@ -105,10 +105,6 @@ input::-webkit-inner-spin-button {
     margin: 0;
 }
 
-.text-field {
-    margin-top: 100px;
-    margin-bottom: 1rem;
-}
 .text-field-input {
     text-align: center;
     display: block;
