--- maze.html
@@ -1,40 +1,0 @@
-<!DOCTYPE html>
-
-
-<html>
-<head>
-    <title>AStar</title>
-    <link rel="stylesheet" href="maze.css">
-</head>
-<body>
-
-<div class="side-border-left"></div>
-<div class=main-part>
-    <div style="display: inline-flex">
-        <h1 class="header-text">
-            A* ALGORITHM
-            <div class="underline"></div>
-        </h1>
-    </div>
-    <div style="text-align: left">
-        <h2 class="interview-right">text 1
-            <div class="underline"></div>
-        </h2>
-        <h2 class="interview-left">text 2
-            <div class="underline"></div>
-        </h2>
-    </div>
-    <div class="interactive-elements">
-        <h2 class="interview-right">Size</h2>
-        <input type="number" id="size"/>
-      <button id="generate">Generate maze</button>
-      <button id="astar">Start algorithm</button>
-    </div>
-    <br><br>
-    <div id="board" class="board"></div>
-    <div class="bottom"></div>
-</div>
-<div class="side-border-right"></div>
-<script src="maze.js"></script>
-</body>
-</html>
